+++ packages/client-p2p/src/handler/blockRequest.js
@@ -3,12 +3,11 @@
 // of the ISC license. See the LICENSE file for details.
 // @flow
 
-import type { BlockRequestMessage } from '../../message/types';
-import type { P2pState } from '../../server/types';
-import type { MessageInterface, PeerInterface } from '../../types';
+import type { BlockRequestMessage } from '../message/types';
+import type { P2pState, MessageInterface, PeerInterface } from '../types';
 
-const message = require('../../message/blockRequest');
-const provideBlocks = require('../../sync/provideBlocks');
+const message = require('../message/blockRequest');
+const provideBlocks = require('../sync/provideBlocks');
 
 module.exports = function handleBlockRequest (self: P2pState, peer: PeerInterface, message: MessageInterface): void {
   self.l.debug(() => [peer.shortId, 'BlockRequest', JSON.stringify(message.encode().message)]);
