+++ packages/client-p2p/src/handler/blockResponse.js
@@ -3,17 +3,16 @@
 // of the ISC license. See the LICENSE file for details.
 // @flow
 
-import type { BlockResponseMessage } from '../../message/types';
-import type { P2pState } from '../../server/types';
-import type { MessageInterface, PeerInterface } from '../../types';
+import type { BlockResponseMessage } from '../message/types';
+import type { P2pState, MessageInterface, PeerInterface } from '../types';
 
-const message = require('../../message/blockResponse');
-const queueBlocks = require('../../sync/queueBlocks');
+const message = require('../message/blockResponse');
+const processBlocks = require('../sync/processBlocks');
 
 module.exports = function handleBlockResponse (self: P2pState, peer: PeerInterface, message: MessageInterface): void {
   self.l.debug(() => [peer.shortId, 'BlockResponse', JSON.stringify(message.encode().message)]);
 
-  queueBlocks(self, peer, (message.raw: BlockResponseMessage));
+  processBlocks(self, peer, (message.raw: BlockResponseMessage));
 };
 
 module.exports.TYPE = message.TYPE;
