+++ packages/client-p2p/src/onPeerDiscovery.js
@@ -3,12 +3,11 @@
 // of the ISC license. See the LICENSE file for details.
 // @flow
 
-import type { PeerInterface } from '../types';
-import type { P2pState } from './types';
+import type { P2pState, PeerInterface } from './types';
 
 const promisify = require('@polkadot/util/promisify');
 
-const defaults = require('../defaults');
+const defaults = require('./defaults');
 
 module.exports = function onPeerDiscovery (self: P2pState): void {
   // flowlint-next-line unclear-type:off
