+++ packages/client-p2p/src/onPeerDiscovery.spec.js
