+++ packages/client-p2p/src/stop.js
