+++ packages/client-p2p/src/sync/peerMax.js
@@ -3,21 +3,19 @@
 // of the ISC license. See the LICENSE file for details.
 // @flow
 
-import type { P2pState } from '../server/types';
-import type { PeerInterface } from '../types';
+import type { P2pState, PeerInterface } from '../types';
 
 const BN = require('bn.js');
 
-const { MAX_SYNC_BLOCKS } = require('../defaults');
 const peerRequests = require('./peerRequests');
 
 module.exports = function peerMax (self: P2pState, peer: PeerInterface): BN {
   const first = self.chain.blocks.getBestNumber().addn(1);
   const requests = peerRequests(self, peer);
 
-  return requests.reduce((next, { peerId, request: { from } }) => {
+  return requests.reduce((next, { request: { from, max } }) => {
     // flowlint-next-line unclear-type:off
-    const last = new BN(((from: any): BN)).addn(MAX_SYNC_BLOCKS);
+    const last = new BN(((from: any): BN)).addn(max);
 
     return last.gt(next)
       ? last
