+++ packages/client-p2p/src/sync/provideBlocks.js
@@ -4,8 +4,7 @@
 // @flow
 
 import type { BlockRequestMessage } from '../message/types';
-import type { P2pState } from '../server/types';
-import type { PeerInterface } from '../types';
+import type { P2pState, PeerInterface } from '../types';
 
 const BN = require('bn.js');
 const isU8a = require('@polkadot/util/is/u8a');
@@ -21,10 +20,11 @@ module.exports = function provideBlocks (self: P2pState, peer: PeerInterface, re
   const blocks = [];
 
   // FIXME: Also send blocks starting with hash
-  let count = isU8a(request.from) ? MAX_SYNC_BLOCKS : 0;
+  const max = Math.min(request.max || MAX_SYNC_BLOCKS, MAX_SYNC_BLOCKS);
+  let count = isU8a(request.from) ? max : 0;
   const increment = request.direction === 'ascending' ? new BN(1) : new BN(-1);
 
-  while (count < MAX_SYNC_BLOCKS && current.lte(best) && !current.isNeg()) {
+  while (count < max && current.lte(best) && !current.isNeg()) {
     const hash = self.chain.state.getBlockHash(current);
 
     blocks.push(
