--- packages/client-p2p/src/sync/queueBlocks.js
@@ -1,11 +1,0 @@
-// Copyright 2017-2018 Jaco Greeff
-// This software may be modified and distributed under the terms
-// of the ISC license. See the LICENSE file for details.
-// @flow
-
-import type { BlockResponseMessage } from '../message/types';
-import type { P2pState } from '../server/types';
-import type { PeerInterface } from '../types';
-
-module.exports = function queueBlocks (self: P2pState, peer: PeerInterface, response: BlockResponseMessage): void {
-};
